pyproject.toml
[tool.poetry]
name = "bot-translator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.26.0"
python-telegram-bot = "^13.7"
loguru = "^0.5.3"
python-dotenv = "^0.19.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
